// project-management-frontend/src/@api/models/admin-analytics-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project Management Tool API
 * API documentation for Project Management Tool
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AdminAnalyticsDto
 */
export interface AdminAnalyticsDto {
    /**
     * 
     * @type {number}
     * @memberof AdminAnalyticsDto
     */
    totalUsers?: number;
    /**
     * 
     * @type {number}
     * @memberof AdminAnalyticsDto
     */
    totalProjects?: number;
    /**
     * 
     * @type {number}
     * @memberof AdminAnalyticsDto
     */
    totalTeams?: number;
    /**
     * 
     * @type {number}
     * @memberof AdminAnalyticsDto
     */
    totalTasks?: number;
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof AdminAnalyticsDto
     */
    tasksByStatus?: { [key: string]: number; };
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof AdminAnalyticsDto
     */
    projectsByStatus?: { [key: string]: number; };
    /**
     * 
     * @type {number}
     * @memberof AdminAnalyticsDto
     */
    overdueTasks?: number;
    /**
     * 
     * @type {number}
     * @memberof AdminAnalyticsDto
     */
    activeUsersLastWeek?: number;
}

/**
 * Check if a given object implements the AdminAnalyticsDto interface.
 */
export function instanceOfAdminAnalyticsDto(value: object): value is AdminAnalyticsDto {
    return true;
}

export function AdminAnalyticsDtoFromJSON(json: any): AdminAnalyticsDto {
    return AdminAnalyticsDtoFromJSONTyped(json, false);
}

export function AdminAnalyticsDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AdminAnalyticsDto {
    if (json == null) {
        return json;
    }
    return {
        
        'totalUsers': json['totalUsers'] == null ? undefined : json['totalUsers'],
        'totalProjects': json['totalProjects'] == null ? undefined : json['totalProjects'],
        'totalTeams': json['totalTeams'] == null ? undefined : json['totalTeams'],
        'totalTasks': json['totalTasks'] == null ? undefined : json['totalTasks'],
        'tasksByStatus': json['tasksByStatus'] == null ? undefined : json['tasksByStatus'],
        'projectsByStatus': json['projectsByStatus'] == null ? undefined : json['projectsByStatus'],
        'overdueTasks': json['overdueTasks'] == null ? undefined : json['overdueTasks'],
        'activeUsersLastWeek': json['activeUsersLastWeek'] == null ? undefined : json['activeUsersLastWeek'],
    };
}

export function AdminAnalyticsDtoToJSON(json: any): AdminAnalyticsDto {
    return AdminAnalyticsDtoToJSONTyped(json, false);
}

export function AdminAnalyticsDtoToJSONTyped(value?: AdminAnalyticsDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'totalUsers': value['totalUsers'],
        'totalProjects': value['totalProjects'],
        'totalTeams': value['totalTeams'],
        'totalTasks': value['totalTasks'],
        'tasksByStatus': value['tasksByStatus'],
        'projectsByStatus': value['projectsByStatus'],
        'overdueTasks': value['overdueTasks'],
        'activeUsersLastWeek': value['activeUsersLastWeek'],
    };
}